+++ test/support/whizbang.ex
@@ -1,42 +1,10 @@
 # credo:disable-for-this-file
-defmodule Whizbang do
-  @moduledoc false
-  use Bonny.Controller
-  use Agent
-
-  def start_link() do
-    Agent.start_link(fn -> [] end, name: __MODULE__)
-  end
-
-  def get() do
-    Agent.get(__MODULE__, fn events -> events end)
-  end
-
-  def get(type) do
-    Agent.get(__MODULE__, fn events ->
-      Keyword.get_values(events, type)
-    end)
-  end
-
-  def put(event) do
-    Agent.update(__MODULE__, fn events -> [event | events] end)
-  end
-
-  @impl true
-  def add(evt), do: put({:added, evt})
-  @impl true
-  def modify(evt), do: put({:modified, evt})
-  @impl true
-  def delete(evt), do: put({:deleted, evt})
-  @impl true
-  def reconcile(evt), do: put({:reconciled, evt})
-end
 
 defmodule V1.Whizbang do
   @moduledoc false
   use Bonny.Controller
-  @kind "Whizzo"
-  @names %{kind: "Whizzo"}
+
+  def conn(), do: Bonny.Test.IntegrationHelper.conn()
 
   @impl true
   def add(_), do: :ok